Completed FriendRequests, Friends, and Announcements.

+++ Backend/mainPackage/src/main/java/mainPackage/friendsService/FriendsController.java
@@ -27,15 +27,15 @@
         private FriendRepository friendRepository;
 
         // Read
-        @GetMapping("/")
+        @GetMapping("/{uid}")
         @Operation(summary = "Find a friend by ID",
                 description = "Returns a friend based on the provided ID.")
         @ApiResponses(value = {
                 @ApiResponse(responseCode = "200", description = "Friend retrieved successfully"),
                 @ApiResponse(responseCode = "404", description = "Friend not found")
         })
-        public GeneralUser findFriendById(@PathVariable(name = "id") int id) {
-            return generalUserRepository.findGeneralUserById(id);
+        public GeneralUser findFriendById(@PathVariable int uid) {
+            return generalUserRepository.findGeneralUserById(uid);
         }
 
         // Delete
@@ -46,15 +46,15 @@
                 @ApiResponse(responseCode = "200", description = "Friend removed successfully"),
                 @ApiResponse(responseCode = "404", description = "Friend not found")
         })
-        public String removeFriend(@PathVariable int uid, @PathVariable int id) {
-            GeneralUser u = generalUserRepository.findGeneralUserById(uid);
-            GeneralUser friendToRemove = generalUserRepository.findGeneralUserById(id);
+        public String removeFriend(@PathVariable int id, @PathVariable int uid) {
+            GeneralUser u = generalUserRepository.findGeneralUserById(id);
+            GeneralUser friendToRemove = generalUserRepository.findGeneralUserById(uid);
 
             if (u == null || friendToRemove == null) {
                 return "Friend or user does not exist.";
             } else {
                 Friend friend = friendRepository.findFriendBySenderAndReceiver(u, friendToRemove);
-                if (friend != null) {
+                if (friend != null && friend.getStatus() == Friend.FriendshipStatus.ACCEPTED) {
                     friendRepository.delete(friend);
                     return "Removed " + friendToRemove.getUserName() + " successfully.";
                 } else {
@@ -76,10 +76,8 @@
             }
 
             List<GeneralUser> friends = new ArrayList<>();
-            List<Friend> userFriends = friendRepository.findFriendRequestsByReceiver(u);
-            for (Friend friend : userFriends) {
-                friends.add(friend.getSender());
-            }
+            friends.addAll(friendRepository.findFriendsBySender(u));
+            friends.addAll(friendRepository.findFriendsByReceiver(u));
             return friends;
         }
     }
